feat: Simplify implementation
There is no need to create a tube with a hollow interior, we can just
simply create the cube and then remove the material to make the hole
in the middle. Later I could make the cube hollow inside and use the
more complex structure, but easier is better for now. The goal is to see
if we can make a "tube" small enough for use a straight pin, which is
about 0.561mm or 561mu in diameter.

The first one failed and this one isn't any different but I want to
simplify the cube so I can eventually make multiple with different sized
tubes.